Refaktoring to avoid duplicate code
