+++ src/main/resources/lib/types/partTypes/statisticFigures.ts
@@ -1,15 +1,9 @@
-import { type AttachmentTablesFiguresData } from '/lib/types/partTypes/attachmentTablesFigures'
+import { type AttachmentTablesFiguresProps } from '/lib/types/partTypes/attachmentTablesFigures'
 
-export interface StatisticFiguresProps {
-  accordions: AttachmentTablesFiguresData[]
-  freeText: string | undefined
-  showAll: string
-  showLess: string
+export interface StatisticFiguresProps extends AttachmentTablesFiguresProps {
   selectedFigures: string
   statbankBoxTitle: string
   statbankBoxText: string
-  appName: string
-  title: string
   icon: string
   iconStatbankBox: string
   statbankHref: string
